% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aconv.R
\name{aconv}
\alias{aconv}
\title{Convolution Tool for Two-Scale Transform.}
\usage{
aconv(f, x)
}
\arguments{
\item{f}{filter.}

\item{x}{1-d signal.}
}
\value{
\code{y} filtered result.
}
\description{
Filtering by periodic convolution of x with the
time-reverse of f.
}
\examples{
qmf <- MakeONFilter('Haar')
x <- MakeSignal('HeaviSine',2^3)
aconv(qmf,x)
}
\seealso{
\code{\link{iconvv}}, \code{\link{UpDyadHi}},
\code{\link{UpDyadLo}}, \code{\link{DownDyadHi}}, \code{\link{DownDyadLo}}.
}
