% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_partition.R
\name{block_partition}
\alias{block_partition}
\title{Construction of the block partition}
\usage{
block_partition(x, L)
}
\arguments{
\item{x}{noisy wc at a given scale.}

\item{L}{block size.}
}
\description{
Construction of the block partition
}
