% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_partition2d.R
\name{block_partition2d}
\alias{block_partition2d}
\title{Construct 2d block partition}
\usage{
block_partition2d(x, L)
}
\arguments{
\item{x}{noisy wc at a given scale.}

\item{L}{block size.}
}
\description{
Construct 2d block partition
}
