% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockThresh.R
\name{BlockThresh}
\alias{BlockThresh}
\title{Wavelet Block thresholding}
\usage{
BlockThresh(wc, j0, hatsigma, L, qmf, thresh = "hard")
}
\arguments{
\item{wc}{wavelet coefficients.}

\item{j0}{coarsest decomposition scale.}

\item{hatsigma}{estimator of noise variance.}

\item{L}{Block size (n mod L must be 0).}

\item{qmf}{Orthonormal quadrature mirror filter.}

\item{thresh}{"hard" or "soft".}
}
\value{
\code{y} Wavelet block thresholding estimator.
}
\description{
Wavelet Block thresholding
}
