% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockThresh2d.R
\name{BlockThresh2d}
\alias{BlockThresh2d}
\title{2d Wavelet Block thresholding}
\usage{
BlockThresh2d(wc, j0, hatsigma, L, qmf, thresh = "soft")
}
\arguments{
\item{wc}{wavelet coefficients.}

\item{j0}{coarsest decomposition scale.}

\item{hatsigma}{estimator of noise variance.}

\item{L}{Block size (n mod L must be 0).}

\item{qmf}{Orthonormal quadrature mirror filter.}

\item{thresh}{'hard' or 'soft'.}
}
\value{
\code{wcb} wavelet coefficient estimators.
}
\description{
This function is used to threshold the coefficients by group (or block).
}
