% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cameraman.R
\docType{data}
\name{cameraman}
\alias{cameraman}
\title{cameraman Image}
\format{
A numeric 512 x 512 matrix.
}
\usage{
data(cameraman)
}
\description{
A dataset containing cameraman image.
}
\keyword{datasets}
