% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircularShift.R
\name{CircularShift}
\alias{CircularShift}
\title{Circular Shifting of a matrix/image}
\usage{
CircularShift(matrix, colshift = 0, rowshift = 0)
}
\arguments{
\item{matrix}{2-d signal (matrix).}

\item{colshift}{column shift index (integer).}

\item{rowshift}{row shift index (integer).}
}
\value{
\code{result} 2-d shifted signal.
}
\description{
Pixels that get shifted off one side of the image
are put back on the other side.
}
\examples{
A <- matrix(1:4, ncol=2, byrow=TRUE)
CircularShift(A, 0, -1)
}
\seealso{
\code{\link{FWT2_TI}}, \code{\link{IWT2_TI}}.
}
