% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownDyadHi.R
\name{DownDyadHi}
\alias{DownDyadHi}
\title{Hi-Pass Downsampling operator (periodized)}
\usage{
DownDyadHi(x, qmf)
}
\arguments{
\item{x}{1-d signal at fine scale.}

\item{qmf}{filter.}
}
\value{
\code{y} 1-d signal at coarse scale.
}
\description{
Hi-Pass Downsampling operator (periodized)
}
\examples{
qmf <- MakeONFilter('Haar')
x <- MakeSignal('HeaviSine',2^3)
DownDyadHi(x, qmf)
}
\seealso{
\code{\link{DownDyadLo}}, \code{\link{UpDyadHi}},
\code{\link{UpDyadLo}}, \code{\link{FWT_PO}}, \code{\link{iconvv}}.
}
