% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownDyadLo.R
\name{DownDyadLo}
\alias{DownDyadLo}
\title{Lo-Pass Downsampling operator (periodized)}
\usage{
DownDyadLo(x, qmf)
}
\arguments{
\item{x}{1-d signal at fine scale.}

\item{qmf}{filter.}
}
\value{
\code{d} 1-d signal at coarse scale.
}
\description{
Lo-Pass Downsampling operator (periodized)
}
\examples{
qmf <- MakeONFilter('Haar')
x <- MakeSignal('HeaviSine',2^3)
DownDyadLo(x,qmf)
}
\seealso{
\code{\link{DownDyadHi}}, \code{\link{UpDyadHi}},
\code{\link{UpDyadLo}}, \code{\link{FWT_PO}}, \code{\link{aconv}}.
}
