% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyad.R
\name{dyad}
\alias{dyad}
\title{Index entire j-th dyad of 1-d wavelet xform}
\usage{
dyad(j)
}
\arguments{
\item{j}{integer.}
}
\value{
\code{ix} list of all indices of wavelet coeffts at j-th level.
}
\description{
Index entire j-th dyad of 1-d wavelet xform
}
\examples{
dyad(0)
}
