% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyadlength.R
\name{dyadlength}
\alias{dyadlength}
\title{Find length and dyadic length of array}
\usage{
dyadlength(x)
}
\arguments{
\item{x}{array of length n = 2^J (hopefully).}
}
\value{
\code{n} length(x).

\code{J} least power of two greater than n.
}
\description{
Find length and dyadic length of array
}
\examples{
x <- MakeSignal('Ramp', 8)
dyadlength(x)
}
\seealso{
\code{\link{quadlength}}, \code{\link{dyad}}
}
