% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWN.R
\name{GWN}
\alias{GWN}
\title{Generation of Gaussian White Noise}
\usage{
GWN(n, sigma)
}
\arguments{
\item{n}{sample size.}

\item{sigma}{standard deviation.}
}
\value{
\code{epsilon} resulting noise.
}
\description{
Generation of Gaussian White Noise
}
\examples{
GWN(10, 0.1)
}
