% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HardThresh.R
\name{HardThresh}
\alias{HardThresh}
\title{Apply Hard Threshold}
\usage{
HardThresh(y, t)
}
\arguments{
\item{y}{Noisy Data.}

\item{t}{Threshold.}
}
\value{
A numeric vector with entries
  \eqn{x_i = y_i \mathbf{1}_{\{|y_i| > t\}}}{x_i = y_i * 1(abs(y_i) > t)}.
}
\description{
Apply Hard Threshold
}
\examples{
f <- MakeSignal('HeaviSine',2^3)
qmf <- MakeONFilter('Daubechies', 10)
L <- 0
wc <- FWT_PO(f, L, qmf)
thr <- 2
wct <- HardThresh(wc, thr)
fhard <- IWT_PO(wct, L, qmf)
}
\seealso{
\code{\link{SoftThresh}}.
}
