% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iconvv.R
\name{iconvv}
\alias{iconvv}
\title{Convolution Tool for Two-Scale Transform.}
\usage{
iconvv(f, x)
}
\arguments{
\item{f}{filter.}

\item{x}{1-d signal.}
}
\value{
\code{y} filtered result.
}
\description{
Filtering by periodic convolution of x with f.
}
\examples{
qmf <- MakeONFilter('Haar')
x <- MakeSignal('HeaviSine',2^3)
iconvv(qmf,x)
}
\seealso{
\code{\link{aconv}}, \code{\link{UpDyadHi}},
\code{\link{UpDyadLo}}, \code{\link{DownDyadHi}}, \code{\link{DownDyadLo}}.
}
