% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSThresh.R
\name{JSThresh}
\alias{JSThresh}
\title{Apply James-Stein Threshold}
\usage{
JSThresh(y, t)
}
\arguments{
\item{y}{Noisy Data.}

\item{t}{Threshold.}
}
\value{
\code{x} filtered result.
}
\description{
(also called the nonnegative garrote)
}
\examples{
f <- MakeSignal('HeaviSine', 2^3)
qmf <- MakeONFilter('Daubechies', 10)
L <- 0
wc <- FWT_PO(f, L, qmf)
thr <- 2
wct <- JSThresh(wc, thr)
fsoft <- IWT_PO(wct, L, qmf)
}
\seealso{
\code{\link{HardThresh}}, \code{\link{SoftThresh}}
}
