% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lshift.R
\name{lshift}
\alias{lshift}
\title{Circular left shift of 1-d signal}
\usage{
lshift(a)
}
\arguments{
\item{a}{1-d signal.}
}
\value{
\code{l} 1-d signal l(i) = x(i+1) except l(n) = x(1).
}
\description{
Circular left shift of 1-d signal
}
\examples{
x <- MakeSignal('HeaviSine',2^3)
lshift(x)
}
