% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeSignal.R
\name{MakeSignal}
\alias{MakeSignal}
\title{Make artificial signal.}
\usage{
MakeSignal(name, n)
}
\arguments{
\item{name}{string, 'HeaviSine', 'Bumps', 'Blocks',
'Doppler', 'Ramp','Cusp', 'Sing'.}

\item{n}{desired signal length}
}
\value{
\code{sig} 1-d signal.
}
\description{
Make artificial signal.
}
\examples{
\dontrun{
sig <- MakeSignal(Name,n)
}
}
\seealso{
\code{\link{FWT_PO}}, \code{\link{IWT_PO}}, \code{\link{FWT2_PO}}, \code{\link{IWT2_PO}}.
}
