% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeSignalNewb.R
\name{MakeSignalNewb}
\alias{MakeSignalNewb}
\title{Make artificial 1-d signal.}
\usage{
MakeSignalNewb(name, n)
}
\arguments{
\item{name}{string, 'Cusp','Step','Wave','Blip','Blocks',
 'Bumps','HeaviSine','Doppler','Angles',
'Parabolas','Time Shifted Sine','Spikes','Corner'}

\item{n}{desired signal length.}
}
\value{
\code{sig} 1-d signal.
}
\description{
Make artificial 1-d signal.
}
\examples{
name <- 'Cusp'
n <- 2^5
sig <- MakeSignalNewb(name,n)
}
\seealso{
\code{\link{FWT_PO}}, \code{\link{IWT_PO}}, \code{\link{FWT2_PO}},
\code{\link{IWT2_PO}}.
}
