% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinMaxThresh.R
\name{MinMaxThresh}
\alias{MinMaxThresh}
\title{Minimax Thresholding}
\usage{
MinMaxThresh(y)
}
\arguments{
\item{y}{signal upon which to perform thresholding.}
}
\value{
\code{x} result.
}
\description{
Minimax Thresholding
}
\references{
D.L. Donoho and I.M. Johnstone (1994). Ideal spatial adaptation by wavelet shrinkage. \emph{Biometrika}, 81(3), 425--455.
}
