% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirrorFilt.R
\name{MirrorFilt}
\alias{MirrorFilt}
\title{Apply (-1)^t modulation}
\usage{
MirrorFilt(x)
}
\arguments{
\item{x}{1-d signal.}
}
\value{
\code{h} 1-d signal with DC frequency content shifted.
                 to Nyquist frequency
}
\description{
h(t) = (-1)^(t-1)  * x(t),  1 <= t <= length(x)
}
\examples{
x <- MakeSignal('HeaviSine',2^3)
h <- MirrorFilt(x)
}
\seealso{
\code{\link{DownDyadHi}}.
}
