% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiMAD.R
\name{MultiMAD}
\alias{MultiMAD}
\title{Apply Shrinkage with level-dependent Noise level estimation}
\usage{
MultiMAD(wc, L)
}
\arguments{
\item{wc}{Wavelet Transform of noisy sequence.}

\item{L}{low-resolution cutoff for Wavelet Transform.}
}
\value{
\code{ws} result of applying VisuThresh to each wavelet level,
                  after scaling so MAD of coefficienst at each level = .6745
}
\description{
Apply Shrinkage with level-dependent Noise level estimation
}
