% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiVisu.R
\name{MultiVisu}
\alias{MultiVisu}
\title{Apply Universal Thresholding to Wavelet Coefficients}
\usage{
MultiVisu(wc, L)
}
\arguments{
\item{wc}{Wavelet Transform of noisy sequence with N(0,1) noise.}

\item{L}{low-frequency cutoff for Wavelet Transform}
}
\value{
\code{x} result of applying VisuThresh to each High Frequency Dyadic Block.
}
\description{
Apply Universal Thresholding to Wavelet Coefficients
}
