% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packet.R
\name{packet}
\alias{packet}
\title{Packet table indexing}
\usage{
packet(d, b, n)
}
\arguments{
\item{d}{depth of splitting in packet decomposition.}

\item{b}{block index among 2^d possibilities at depth d.}

\item{n}{length of signal.}
}
\value{
\code{p} linear indices of all coeff's in that block.
}
\description{
Packet table indexing
}
\examples{
packet(1, 1, 8)
}
