% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSpikes.R
\name{PlotSpikes}
\alias{PlotSpikes}
\title{Plot 1-d signal as baseline with series of spikes.}
\usage{
PlotSpikes(base, t, x, L, J)
}
\arguments{
\item{base}{number, baseline level.}

\item{t}{ordinate values.}

\item{x}{1-d signal, specifies spike deflections from baseline.}

\item{L}{level of coarsest scale.}

\item{J}{least power of two greater than n.}
}
\value{
A plot of spikes on a baseline.
}
\description{
Plot 1-d signal as baseline with series of spikes.
}
\examples{
\dontrun{
PlotSpikes(base,t,x, L, J)
}
}
\seealso{
\code{\link{PlotWaveCoeff}}.
}
