% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotWaveCoeff.R
\name{PlotWaveCoeff}
\alias{PlotWaveCoeff}
\title{Spike-plot display of wavelet coefficients.}
\usage{
PlotWaveCoeff(wc, L, scal)
}
\arguments{
\item{wc}{1-d wavelet transform.}

\item{L}{level of coarsest scale.}

\item{scal}{scale factor (0 ==> autoscale).}
}
\value{
A display of wavelet coefficients (coarsest level
        NOT included) by level and position.
}
\description{
Spike-plot display of wavelet coefficients.
}
\examples{
\dontrun{
PlotWaveCoeff(wc,L,scal)
}
}
\seealso{
\code{\link{FWT_PO}}, \code{\link{IWT_PO}}, \code{\link{PlotSpikes}}.
}
