% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadlength.R
\name{quadlength}
\alias{quadlength}
\title{Find length and dyadic length of square matrix}
\usage{
quadlength(x)
}
\arguments{
\item{x}{2-d image; dim(n,n), n = 2^J (hopefully).}
}
\value{
\code{n} length(x).

\code{J} least power of two greater than n.
}
\description{
h(t) = (-1)^(t-1)  * x(t),  1 <= t <= length(x)
}
\examples{
quadlength(matrix(1:16,ncol=4))
}
