% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaphNMR.R
\docType{data}
\name{RaphNMR}
\alias{RaphNMR}
\title{Nuclear magnetic resonance (NMR) signal}
\format{A numeric vector of length 1024.}
\source{
MRS Unit, VA Medical Center, San Francisco. Adrain Maudsley, Ph.D., Professor of Radiology.
    This NMR signal was obtained from Chris Raphael, then a
    postdoctoral fellow in the Department of Statistics at
    Stanford University who was working on Hidden Markov Models
    for restoring NMR Spectra.
}
\usage{
data(RaphNMR)
}
\description{
A dataset containing a NMR signal.
}
\keyword{datasets}
