% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repmat.R
\name{repmat}
\alias{repmat}
\title{Replicate and tile an array.}
\usage{
repmat(a, n, m)
}
\arguments{
\item{a}{input array (scalar, vector, matrix)}

\item{n}{number of time to repeat input array in row and column dimensions}

\item{m}{repetition factor}
}
\description{
Replicate and tile an array.
}
