% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rshift.R
\name{rshift}
\alias{rshift}
\title{Circular right shift of 1-d signal}
\usage{
rshift(a)
}
\arguments{
\item{a}{1-d signal.}
}
\value{
\code{r} 1-d signal r(i) = x(i-1) except r(1) = x(n).
}
\description{
Circular right shift of 1-d signal
}
\examples{
x <- MakeSignal('HeaviSine', 2^3)
rshift(x)
}
