% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeAsRow.R
\name{ShapeAsRow}
\alias{ShapeAsRow}
\title{Make signal a row vector}
\usage{
ShapeAsRow(sig)
}
\arguments{
\item{sig}{a row or column vector.}
}
\value{
\code{row} a row vector.
}
\description{
Make signal a row vector
}
\examples{
sig <- matrix(1:4)
row <- ShapeAsRow(sig)
}
