% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLphantom.R
\docType{data}
\name{SLphantom}
\alias{SLphantom}
\title{3-d Shepp-Logan phantom}
\format{A numeric array of size 64x64x64.}
\usage{
data(SLphantom)
}
\description{
A dataset containing a 3d head phantom that can be used
to test 3-d reconstruction algorithms.
Shepp-Logan phantom is well-known imitation of human cerebral.
}
\keyword{datasets}
