% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUREThresh.R
\name{SUREThresh}
\alias{SUREThresh}
\title{Adaptive Threshold Selection Using Principle of SURE}
\usage{
SUREThresh(y)
}
\value{
y  Noisy Data with Std. Deviation = 1.

thresh Threshold used.
}
\description{
SURE referes to Stein's Unbiased Risk Estimate.
}
