% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpDyadHi.R
\name{UpDyadHi}
\alias{UpDyadHi}
\title{Hi-Pass Upsampling operator; periodized}
\usage{
UpDyadHi(x, qmf)
}
\arguments{
\item{x}{1-d signal at coarser scale.}

\item{qmf}{filter.}
}
\value{
\code{u} 1-d signal at finer scale.
}
\description{
Hi-Pass Upsampling operator; periodized
}
\examples{
qmf <- MakeONFilter('Haar')
x <- MakeSignal('HeaviSine',2^3)
UpDyadHi(x,qmf)
}
\seealso{
\code{\link{DownDyadLo}}, \code{\link{DownDyadHi}},
\code{\link{UpDyadLo}}, \code{\link{IWT_PO}}, \code{\link{aconv}}.
}
