% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpDyadLo.R
\name{UpDyadLo}
\alias{UpDyadLo}
\title{Lo-Pass Upsampling operator; periodized}
\usage{
UpDyadLo(x, qmf)
}
\arguments{
\item{x}{1-d signal at coarser scale.}

\item{qmf}{filter.}
}
\value{
\code{y} 1-d signal at finer scale.
}
\description{
Lo-Pass Upsampling operator; periodized
}
\examples{
qmf <- MakeONFilter('Haar')
x <- MakeSignal('HeaviSine',2^3)
UpDyadLo(x,qmf)
}
\seealso{
\code{\link{DownDyadLo}}, \code{\link{DownDyadHi}},
\code{\link{UpDyadHi}}, \code{\link{IWT_PO}}, \code{\link{iconvv}}.
}
