% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpSampleN.R
\name{UpSampleN}
\alias{UpSampleN}
\title{Upsampling operator}
\usage{
UpSampleN(x, s)
}
\arguments{
\item{x}{1-d signal, of length n.}

\item{s}{upsampling scale, default = 2.}
}
\value{
\code{y} 1-d signal, of length s*n with zeros
                 interpolating alternate samples
                 y(s*i-1) = x(i), i=1,...,n
}
\description{
Upsampling operator
}
