% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValSUREThresh.R
\name{ValSUREThresh}
\alias{ValSUREThresh}
\title{Adaptive Threshold Selection Using Principle of SURE}
\usage{
ValSUREThresh(x)
}
\arguments{
\item{x}{Noisy Data with Std. Deviation = 1.}
}
\value{
\code{thresh} Value of Threshold.
}
\description{
SURE referes to Stein's Unbiased Risk Estimate.
}
