% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisuThresh.R
\name{VisuThresh}
\alias{VisuThresh}
\title{Visually calibrated Adaptive Smoothing}
\usage{
VisuThresh(y, thresh = "soft")
}
\arguments{
\item{y}{Signal upon which to perform visually calibrated Adaptive Smoothing.}

\item{thresh}{'hard' or 'soft'.}
}
\value{
\code{x} result of applying VisuThresh.
}
\description{
Visually calibrated Adaptive Smoothing
}
\references{
D.L. Donoho and I.M. Johnstone (1994). Ideal spatial adaptation by wavelet shrinkage. \emph{Biometrika}, 81(3), 425--455.
}
