% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveFEX.R
\name{WaveFEX}
\alias{WaveFEX}
\title{Wavelet-based Feature EXtraction}
\usage{
WaveFEX(type = "PO", obs, qmf, scale = log2(nrow(obs)), iwt = F, alpha = F)
}
\arguments{
\item{type}{Orthogonal transform: PO or translation invariant transform:TI.}

\item{obs}{A matrix of n x p functional data (n should be a power of 2).}

\item{qmf}{Orthonormal quadrature mirror filter.}

\item{scale}{Finest scale of the decomposition.}

\item{iwt}{Logical to specify forward (FALSE) or inverse (TRUE) wavelet transform.}

\item{alpha}{Logical TRUE to include scaling coefficients at the first scale.}
}
\value{
\code{contj} Matrix  scale x p of Squared norm of the wavelet coefficient at each scale.

\code{wc} wavelet coefficients matrix.
}
\description{
Todo: add description
}
\references{
Antoniadis, A., Brossat, X., Cugliari, J., & Poggi, J. M. (2013). Clustering functional data using wavelets. \emph{Int. J. Wavelets Multiresolution Inf. Process.}, 11(01), 1350003.

Cheam, A., Fredette, M., Marbac, M., & Navarro, F. (2023). Translation-invariant functional clustering on COVID-19 deaths adjusted on population risk factors. \emph{J. R. Stat. Soc., C: Appl. Stat.}, 72(2), 387-413.
}
