% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubelength.R
\name{cubelength}
\alias{cubelength}
\title{Find length and dyadic length of square array}
\usage{
cubelength(x)
}
\arguments{
\item{x}{3-d array; dim(n,n,n), n = 2^J (hopefully).}
}
\value{
\code{n} length(x).

\code{J} least power of two greater than n.
}
\description{
3d counterpart of Donoho's quadlength utilized by the 2d pair.
Original matlab code Vicki Yang and Brani Vidakovic.
}
\examples{
cubelength(array(1:3, c(2,2,2)))
}
\seealso{
\code{\link{FWT3_PO}}, \code{\link{IWT3_PO}}.
}
