% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVlinear.R
\name{CVlinear}
\alias{CVlinear}
\title{2-Fold Cross Validation for linear estimator}
\usage{
CVlinear(Y, L, qmf, D, wc)
}
\arguments{
\item{Y}{Noisy observations.}

\item{L}{Level of coarsest scale.}

\item{qmf}{Orthonormal quadrature mirror filter.}

\item{D}{Dimension vector of the models considered.}

\item{wc}{1-d wavelet coefficients.}
}
\value{
\code{CritCV} Cross validation criteria.

\code{hat_f_m_2FCV}
}
\description{
Selection of the number of wavelet coefficients to be maintained by the
cross validation method proposed by Nason in the case of threshold
selection. This method is adapted here to select among linear estimators.
}
\references{
Nason, G. P. (1996). Wavelet shrinkage using cross-validation. \emph{Journal of the Royal Statistical Society: Series B}, 58(2), 463--479.

Navarro, F. and Saumard, A. (2017). Slope heuristics and V-Fold model selection in heteroscedastic regression using strongly localized bases. \emph{ESAIM: Probability and Statistics}, 21, 412--451.
}
