% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invblock_partition.R
\name{invblock_partition}
\alias{invblock_partition}
\title{Inversion of the 1d block partition}
\usage{
invblock_partition(x, n, L)
}
\arguments{
\item{x}{noisy wc at a given scale.}

\item{n}{scale.}

\item{L}{block size.}
}
\description{
Inversion of the 1d block partition
}
