% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invblock_partition2d.R
\name{invblock_partition2d}
\alias{invblock_partition2d}
\title{Inversion of the 2d block partition}
\usage{
invblock_partition2d(x, n, L)
}
\arguments{
\item{x}{noisy wc at a given scale.}

\item{n}{scale.}

\item{L}{block size.}
}
\description{
Inversion of the 2d block partition
}
