% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNR.R
\name{SNR}
\alias{SNR}
\title{Signal/Noise ratio}
\usage{
SNR(x, y)
}
\arguments{
\item{x}{Original reference signal.}

\item{y}{Restored or noisy signal.}
}
\value{
Signal/Noise ratio.
}
\description{
Signal/Noise ratio
}
\examples{
n <- 2^4
x <- MakeSignal('HeaviSine', n)
y <- x + rnorm(n, mean=0, sd=1)
SNR(x, y)
}
