% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUREThresh.R
\name{SUREThresh}
\alias{SUREThresh}
\title{Adaptive Threshold Selection Using Principle of SURE}
\usage{
SUREThresh(y)
}
\arguments{
\item{y}{Noisy Data with Std. Deviation = 1.}
}
\value{
\code{x} Estimate of mean vector

\code{thresh} Threshold used.
}
\description{
SURE referes to Stein's Unbiased Risk Estimate.
}
